+++ CHANGELOG.md
@@ -7,10 +7,14 @@ This project uses [Semantic Versioning](https://semver.org/).
 ---
 ## [0.5.0] - 2026-06-10 "Längere Haltezeit am Bahnhof"
 
-### Changed
+### Added
 
 - **Being replanned tag** - `VORLAEUFIG_NICHT_REKONSTRUIERBAR` new status tag added
 - **Delete cache button for saved trips in past view** - allows to delete saved trips that have been reconstructed from cache in the past view mimicking the option of bahn.de to delete saved trips, only affects cache
+
+### Fixed
+
+- **indentation** - thoughout
 
 ---