+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project uses [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
-- **indentation** - thoughout
+- **indentation** - throughout
 
 ---